THESIS_VITALIOT_WEBBV2/requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
crypto==1.4.1
cryptography==2.8
csvwriter==0.2.2
cycler==0.10.0
decorator==4.4.0
Django==2.2.1
django-crispy-forms==1.7.2
docopt==0.6.2
entrypoints==0.3
firebase==3.0.1
firebase-admin==3.0.0
flake8==3.7.8
future==0.18.1
gcloud==0.17.0
geocoder==1.38.1
get==2019.4.13
google-api-core==1.14.2
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.23.0
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
imageio==2.5.0
instaloader==4.2.8
isort==4.3.21
joblib==0.13.2
jwcrypto==0.6.0
jws==0.1.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.0
mccabe==0.6.1
msgpack==0.6.1
Naked==0.1.31
networkx==2.3
numpy==1.16.4
oauth2client==3.0.0
oauthlib==3.0.2
opencv-python==4.1.2.30
pandas==0.25.0
patsy==0.5.1
pefile==2019.4.18
Pillow==6.0.0
post==2019.4.13
protobuf==3.9.1
public==2019.4.13
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.4.3
pycryptodomex==3.9.0
pyflakes==2.1.1
PyInstaller==3.5
pylint==2.3.1
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
Pyrebase==3.0.27
pyserial==3.4
python-dateutil==2.8.0
python-firebase==1.2
python-jwt==2.0.1
pytz==2019.1
PyWavelets==1.0.3
pywin32-ctypes==0.2.0
PyYAML==5.1.2
pyzbar==0.1.8
query-string==2019.4.13
ratelim==0.1.6
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
shellescape==3.4.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
sseclient==0.0.24
statsmodels==0.10.1
thingspeak==1.0.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.9
whitenoise==4.1.4
wrapt==1.11.2
XlsxWriter==1.1.8
